Rework keyboard controls
The third argument is the "root" note, the lowest one available. On
en_US keyboard you would play it by pressing "Q".

The keyboard layout was rearranged. Instead of keycode = MIDI note, the
letter keys (and punctuation) now correspond to a number _added_ to the
root note. Since "Q" plays the lowest note, its number is just 0. "A"
plays a note one octave higher, so its number is 12. "]" plays the note
before that, so 11; etc.

Forking leaves MIDI ports behind (why?), so it was removed.